Changes to test using a helper function for view switching
